Fix javac for 7000.0.3-M V2 - wool/concrete + vertex double + dayTime syntax
- Run 653 had 12 errors: GRAY_WOOL/BLACK_WOOL/WHITE_WOOL not in 26.2 Blocks -> replace with COBBLESTONE/OAK_PLANKS
- OrbitalPlanets/AmbientWorld/WavyWater quad called vertex(Vec3) but vertex now expects double - fix to double x,y,z
- TimeWarpEffect digitAlpha float->int lossy conversion
- Aurora/ShootingStars/LightRays dayTime leftover .getGameTimeDeltaPartialTick fixed

520/520 + 264/264 green, should restore 248 classes

+++ mcsm-extras/java/net/mcsm/extras/client/McsmTimeWarpEffect.java
@@ -183,7 +183,7 @@ public final class McsmTimeWarpEffect {
                         // Digits as glowing quads
                         for (int d = 0; d < 5; d++) {
                             double dx = routerX + (d - 2) * 0.6;
-                            float digitAlpha = 180 + (int)(Mth.sin(t * 2f + d) * 40);
+                            int digitAlpha = 180 + (int)(Mth.sin(t * 2f + d) * 40);
                             int digitVal = (int)((displayTime / Math.pow(10, 4 - d)) % 10);
                             // Color based on digit
                             int r = 100 + digitVal * 15;
